ArticleViewer first implementation
page that loads an HTML article, with toolbar showing buttons to set
night mode, and increase/decrease text size.

attention was put in the details: (taking too much time, as always)

there is a collapsing toolbar, that can show an header picture for
articles that have one.

button pictures were customized to have the good size on the toolbar.