SimpleMagToc class, to create the summary.
these data class contain the list of downloaded articles, ordered by
categories, with their names and the path to the file on storage.

It also contains the Edito.

The implementation had to be made dirty with the SimpleMagTocExpandable,
as I needed the class to inherit a class for the UI, but this prevented
the serialization of the class for backup.

So SimpleMagToc is used at magasine creation to get it's summary,
it is retreived when launching MagSummary activity, and a
SimpleMagTocExpandable is created from this, to set up the UI.
complicated!